a lot of changes

+++ src/main/java/pl/edu/pk/cosmo/habsatbackend/repository/DataRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pl.edu.pk.cosmo.habsatbackend.entity.Data;
 
 @Repository
-public interface DataRepository extends JpaRepository<Data, Integer> {
+public interface DataRepository extends JpaRepository<Data, Long> {
 }
